Move files around, fix small issues
