+++ examples/AlwaysOnTraceExample.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 require __DIR__ . '/../vendor/autoload.php';
 
 use OpenTelemetry\Sdk\Trace\Attributes;
-use OpenTelemetry\Sdk\Trace\Sampling\AlwaysOnSampler;
-use OpenTelemetry\Sdk\Trace\Sampling\SamplingResult;
+use OpenTelemetry\Sdk\Trace\AlwaysOnSampler;
+use OpenTelemetry\Sdk\Trace\SamplingResult;
 use OpenTelemetry\Sdk\Trace\SimpleSpanProcessor;
 use OpenTelemetry\Sdk\Trace\TracerProvider;
 use OpenTelemetry\Sdk\Trace\ZipkinExporter;
