+++ sdk/Tests/AlwaysParentSamplerTest.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry\Tests\Unit\Tracing\Sampling;
+namespace OpenTelemetry\Sdk\Tests;
 
-use OpenTelemetry\Sdk\Trace\Sampling\AlwaysOnSampler;
-use OpenTelemetry\Sdk\Trace\Sampling\AlwaysParentSampler;
-use OpenTelemetry\Sdk\Trace\Sampling\SamplingResult;
+use OpenTelemetry\Sdk\Trace\AlwaysOnSampler;
+use OpenTelemetry\Sdk\Trace\AlwaysParentSampler;
+use OpenTelemetry\Sdk\Trace\SamplingResult;
 use OpenTelemetry\Sdk\Trace\SpanContext;
 use PHPUnit\Framework\TestCase;
