+++ sdk/Trace/AlwaysOnSampler.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry\Sdk\Trace\Sampling;
+namespace OpenTelemetry\Sdk\Trace;
 
 use OpenTelemetry\Trace as API;
 
@@ -10,7 +10,7 @@ use OpenTelemetry\Trace as API;
  * This implementation of the SamplerInterface always records.
  * Example:
  * ```
- * use OpenTelemetry\Sdk\Trace\Sampler\AlwaysOnSampler;
+ * use OpenTelemetry\Sdk\Trace\AlwaysOnSampler;
  * $sampler = new AlwaysOnSampler();
  * ```
  */
@@ -26,9 +26,10 @@ class AlwaysOnSampler implements Sampler
         string $spanId,
         string $spanName,
         // API\SpanKind $spanKind,
-        array $attributes = [],
-        array $links = []
+        ?API\Attributes $attributes = null,
+        ?API\Links $links = null
     ): SamplingResult {
+        // todo: hook up $attributes and $links
         return new SamplingResult(SamplingResult::RECORD_AND_SAMPLED);
     }
