+++ sdk/Trace/ProbabilitySampler.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry\Sdk\Trace\Sampling;
+namespace OpenTelemetry\Sdk\Trace;
 
 use OpenTelemetry\Trace as API;
 
@@ -10,7 +10,7 @@ use OpenTelemetry\Trace as API;
  * This implementation of the SamplerInterface records with given probability.
  * Example:
  * ```
- * use OpenTelemetry\Trace\Sampler\ProbabilitySampler;
+ * use OpenTelemetry\Trace\ProbabilitySampler;
  * $sampler = new ProbabilitySampler(0.01);
  * ```
  */
@@ -40,8 +40,8 @@ class ProbabilitySampler implements Sampler
         string $spanId,
         string $spanName,
         // API\SpanKind $spanKind,
-        array $attributes = [],
-        array $links = []
+        ?API\Attributes $attributes = null,
+        ?API\Links $links = null
     ): SamplingResult {
         if (null !== $parentContext && ($parentContext->getTraceFlags() & API\SpanContext::TRACE_FLAG_SAMPLED)) {
             return new SamplingResult(SamplingResult::RECORD_AND_SAMPLED);
